Replaced yacc/bison parser for .efg files with a C++ parser. Should make future extension of .efg files simpler.
Notes:
(1) Not 100% tested, but read-write-read-write cycles produce
the same output file;
(2) This version does not enforce that action labels and/or
outcome values are the same when the same ones appear more
than once in the file.  Instead, the first-seen ones are
used.  Shouldn't be a problem; however, probably should add
an error check;
(3) Error reporting should be better than it is.  (It's no
worse than in the .yy files, but having gotten rid of the
yacc/bison parser, we should be able to do better.)