Make bounds-checking functions in RectArray public.
Visual Studio 2005 did not like having the bounds-checking functions
in RectArray being protected, since it would not allow friends of
derived classes to use them. While this is really a bug in VS2005,
on further analysis, there is no real reason not to allow these functions
to be public anyway.