Bugfix to remove duplicate exception name in LCP solver.
The exception class EquilibriumLimitReached appeared in both
efglcp.cc and nfglcp.cc.  Both were in anonymous namespaces,
which should have avoided a name clash.  However, gcc-4.2 appears
not to implement this, or not implement it correctly.

This changes the classes to have different names to explicitly
disambiguate between them.

The EquilibriumLimitReached concept may be more generally useful,
and the exception likely should be refactored into a more visible
location.