Refactor automatic creation of outcomes into GameRep ctor from file.
Add optional parameter p_sparseOutcomes to GameRep strategic form table ctor.
Previous behavior was that strategic form tables were initialized to have
outcomes set to null for all contingencies.  For the typical application,
this is inconvenient and/or inefficient; in reading large games, for instance,
it was found that pre-creating all the outcomes was much, much faster.  In
addition, when operating programmatically, having to create outcomes before
being able to set payoffs is a major bother.

Therefore, p_sparseOutcomes now defaults to false, in which case the
table of contingencies is pre-populated with outcomes.  The old behavior
is still available by specifying p_sparseOutcomes = true; this may be
useful in the odd case in which the majority of outcomes are going to
have zero payoff, or in which the number of outcomes is going to be much
smaller than the number of contingencies.

Remove on-the-fly creation of outcomes from payoff body nfg reader.