Bugfix for local likelihood optima in QRE maximum likelihood estimation.
The maximum likelihood mode of gambit-logit was terminating
as soon as it found an interior local extremum of the likelihood
function along the branch.  This is not correct; there may be
multiple local maxima, and/or the extremum encountered may actually
be a minimizer, especially if the data do not conform closely to the
predictions along the QRE branch.

This modifies the path following code to permit easy restart, by
returning the last point computed before switching to the Newton
search to find the extremum.  In maximum likelihood mode, the
program automatically restarts at this point after polishing the
extremum, to look for other potential interior extrema, or to handle
the possibility that the limiting Nash equilibrium may also be the
global maximizer of the likelihood function.

Fix to deal with possibility of local likelihood optima.