Added tests to verify add player functionality.
This adds tests for the add() method for the players collection in a game. In the case of a strategic game it will add a player with 1 strategy and in the case of an extensive game it will add a player with no moves.

Additionally, the python api documentation was updated to reflect the add player functionality.