Added additional functionality for the infoset class.
Specifically, this commit adds a function to retrieve the player associated with the infoset as well as a test to see if a particular game node precedes the infoset in the game tree.

The class was added to the documentation and the methods and properties were added. Additionally, tests were created to validate the use of the infoset object. Part of the tests involved including an extensive form game derived form the gambit supplied games.