Implement StrategySupport in the Python API
This commit adds full manipulation of strategy supports in the
Python API. All relevant game objects (players, strategies and
outcomes) now can be defined from supports seamlessly and behave
as on a game, intermingling data from both as needed.
All objects defined on supports have the 'unrestrict()' method,
which returns the original version of the object as defined on the
base game.