Implement StrategySupportProfile, StrategicRestriction in Python.
Implement new concepts pertaining to manipulation of sets of strategies.

A StrategySupportProfile is a smart subset of strategies, similar to the
current StrategySupport in C++.  It maintains the requirement that at all
times, each player has at least one strategy in the support.

A StrategicRestriction is an immutable view on another game, defined by
restricting the set of strategies in the original game.  There is no
similarly-encapsulated concept in the C++ as yet.