Implement wxWidgets 2.9.4 compatibility in graphical interface.
This revises the graphical interface implementation to support
wxWidgets 2.9.4.

wxWidgets 2.8.12 compatibility is retained.  Notably, this requires
that dcsvg.(cc,h) remain; this is available in wxWidgets 2.9.4
but not in 2.8.12.  The 2.9.4 build uses the wxWidgets implementation.

There is a bug involving drag-and-drop on OS X with wxWidgets 2.9.4
This requires using 2.8.12 for OS X builds, which requires some
extra steps when building on Snow Leopard.  These are now documented
in the manual.