Extend command-line tools to accept file name on command line.
Extends the tools so that the file name may be specified on the command
line, instead of being passed as standard input.

The previous behavior of reading from standard input is preserved in
the case in which no file is given on the command line, so this is
fully backwards-compatible.