Change Python strategy support representation internally to use pointer.
Modifies the internal representation of strategy supports and strategic
restrictions to use a pointer to the C++ StrategySupport object.
This is occasioned by a change between Cython 0.17 and 0.17.3 which makes
storing the object directly more difficult (owing to the lack of a
sensible default ctor for the C++ class).

This change provides appropriate __init__ and __dealloc__ logic in the
Cython implementation to ensure allocation is correct.  All tests
are passed.  (And the tests were indeed useful in shaking out the
memory management.)