Improve handling of invalidated objects in C++ and Python
Corrects and improves dealing with invalidated objects.

(1) At the C++ level, the dereference operator of GameObjects now
distinguishes between the case of a null pointer versus the case of an
invalidated object (one which has been deleted from its game).
A separate exception is raised.

(2) In Python, the corresponding derefernce operator now catches both
these exceptions and turns them into RuntimeError exceptions.  Previously,
the exception was going uncaught (and leading to a crash of the Python
binary).