Implement getting strategic contingency for game in extensive form
The C++ implementation of game representation does not implement
a concept of an outcome in a strategic game derived from an extensive
game.

This commit provides an implementation at the Python (Cython) level
of a shim class TreeGameOutcome which serves this function, so that
the subscript operation on games works as expected for navigating
the strategic game derived from an extensive game.

Editing operations are not supported by this object as they do not
make sense in this context.