Correct sizing of action editor in move dialog in wxWidgets 2.9.5.
In wxWidgets 2.9.5, sizing of grids and sheets works a bit differently.
This adds an explicit invalidation of the previous best size of the
action editor grid to force recalculation when passed to a sizer.
A more robust fix would correct this in wxSheet, as from various
posts it appears this has to do with size calculation not being invalidated
when grids are created or resized.

Along the way this refactors the action editor sheet class for improved
encapsulation of the implementation.