Bug fix in tracing strategic QRE on non-full supports.
The calcuation of function evaluations and Jacobians in strategic
QRE were incorrect in the case when the strategy profile does not
have full support.

In addition to correcting this bug, this introduces an implementation
of a player relative to a strategy support, which matches in part the
interface to the general GamePlayerRep.  This anticipates the eventual
introduction of strategic restrictions of a game, which currently are
only implemented in Python.