Fix bug in reference counting pure strategy profiles.
PureStrategyProfile had been reimplemented as a std::auto_ptr to an
internal representation class.  This was not correct vis-a-vis how this
class is used, as assigning a profile to another profile is assumed
to have copy semantics.  With an auto_ptr, these assignments were instead
transfering the internal representation pointer, leading to null
pointer dereferencing errors.  This was primarily observed in the
graphical interface, when editing strategic games after adding
strategies, but could also be triggered elsewhere.