Revise gambit-enumpure to require explicit flag for agent equilibria.
The behavior of gambit-enumpure previously was to report agent
equilibria for extensive games when the -S flag was not included.
This was often confusing to users, especially when using the graphical
interface.

This changes the behavior so that now the default is to use strategic
form strategies.  When the -S flag is present, the output is in
strategic form strategies; when not, strategies are converted back to
behavior strategies.

The previous behavior of returning agent form strategies is now
triggered by a new command-line switch -A.

The effect on the graphical interface will be that computing
pure strategy equilibria on the extensive game will return what most
users typically would expect.  Computing all agent equilibria is now
not accessible from the graphical interface, but this is an advanced
concept anyway and something only quite specialist users would require.