Code tidiness updates for clang LLVM 5.1 on OS X 10.9.
This commit bundles a number of mostly minor revisions to the code
to deal with deprecations and compiler warnings which arise starting
with the LLVM 5.1 clang compiler on OS X 10.9.

Mostly these are simply removing unused variables.  There are a few
rewrites of casts in Pelican; as Pelican is so old and dubiously
written in the first place, there is no reason to believe this makes
Pelican any less reliable than it already is.