Implementation of simple STL-like forward iterators on Array and List.
This provides some simple STL-like forward iterator classes, both
const and non-const, for the Array and List classes.

It is recommended that new code use the STL-like features of these
classes, and that existing code be retrofitted as appropriate.
Additional STL-compatible operations on these classes should be
added as required.