Implement normalize operation on mixed profiles.
Mixed strategy and behavior profiles have never been required to sum to
one, as the entries in the vectors can usefully be interpreted as, for
example, counts rather than probabilities.  However, normalization of
a profile such that each player (or information) set's entries summed to
one, to be a proper probability distribution, has never been implemented.

This commit implements a Normalize() (in C++) and normalize() (in Python)
operation, which does just this. For any player or information sets where
the sum of entries is non-zero, the entries are re-scaled to sum to one.
In the case the sum is zero, the entries are left at zero.