Wrap strategic game LCP solver in Python.
This is a preliminary wrapping of the strategic game LCP solver class
in Python.  This is the first Nash equilibrium solver to be accessible
directly in the Python process (as opposed to using external calls to
the command-line tools).

This should be viewed as a preliminary and provisional implementation;
interfaces may be adjusted and the organization of the code may
be able to be improved.