Removes references to profiles on non-full supports from simpdiv solver.
This removes references to mixed strategy profiles with possibly non-full
supports from the simplicial subdivision solver.  Equivalent functionality
can now be obtained using the restricted game mechanism.

In carrying out this removal, adjacent parts of the implementation have
been reformatted and cleaned up per more modern coding standards where
it was straightforward to do so, although not so much as to risk
introducing bugs which would have nothing to do with the removal of
the support-based features.  Conversely, the fact that parts of this
implementation are unchanged does not mean that reformatting,
restructuring, or refactoring of those parts should not be done!  This
remains very old code in substantial need of updating.