Create iterator over strategy support profile.
Implements a StrategySupport::iterator class, migrating and refactoring
what was previously called StrategyCursorForSupport in gambit-enumpoly.

Cleans out the support generation implementation in gambit-enumpoly for
strategic games, removing several unused routines, and modifying the
key one (PossibleNashSubsupports) to use the new iterator.

PossibleNashSubsupports is a candidate in future for a new class of
solver algorithm classes, which return supports or lists of supports
(instead of strategy profiles, as Nash solvers do).