Refactor gambit-simpdiv Nash solver to be derived from Nash solver class.
This refactors the implementation of gambit-simpdiv so it is expressed
as a subclass of the Nash solver class.

This is old and untidy code.  As such, this commit is largely about
making the public interface to the routine compatible with the Nash
solver framework.  The internals remain less than ideal.