Implement fuller access to LCP solver library in Python.
This builds out a fuller implementation of the Python interface to the
LCP Nash equilibrium methods.

A new convenience function, gambit.nash.lcp_solve(), is provided for
one-line access to computing equilibria, where the user does not need
to worry about customizing behavior of the algorithm.

Wrappers for individual algorithm classes for LCP solvers are now
also provided.