Fix bug in QRE maximum likelihood with bifurcations.
When a game has a bifurcation in its QRE correspondence, the orientation
of the path of QRE flips over, meaning the tangent just after the
bifurcation is (approximately) -1 times the tangent before.  This
confuses the maximum likelihood estimation, as well as it would any
application which uses the derivative information.

This fixes the bug by ensuring that comparisions between tangents at
adjacent steps are done with the tangents oriented in the same sign
convention.