Clean up references to abs().
Various calls to abs() (to compute absolute values) appear throughout
the codebase.  These have relied on overloading to resolve arguments
correctly, including the defintion of an inline abs(double) in the
global namespace.  This creates clashes with some compilers.
Therefore, this commit defines an abs(double) in the Gambit namespace,
and all calls reference it, for the avoidance of ambiguity.