Print equilibria on the fly when using enummixed -L
On-the-fly output of equilbria was disabled a few commits ago
due to an unresolved performance issue.  It has since been
established that the performance issue is either specific to
one compiler (clang-703.0.31 on OS X 10.11) or some sort of
fluke.  Therefore, this commit reinstates on-the-fly
equilibrium reporting, although the source of the performance
problem remains uncertain.