Change default starting profile in simpdiv.
This sets the starting profile in simpdiv to the mixed strategy profile
in which each player places all the probability on their first strategy.

This should tend to make it easy for simpdiv to find equilibria in which
players use strategies with probabilities with small denominators, and
so is a not-unreasonable default.