Implement optional disabling of game tree canonicalization.
This implements an optional feature to disable game tree canonicalization
after every change.   For game trees of reasonable size, profiling revealed
just about all the time was spent re-ordering and re-numbering elements
of the game after each new node was added.

There is now a SetCanonicalization() member of GameTreeRep which
permits disabling this.  It is used by the .efg file reader.
The implications are huge: for one example game with around 100k nodes,
loading time went from 6 minutes to 2 seconds.

Canonicalization is largely a housekeeping feature, as no algorithm
depends on the numbering.  It might be worth considering more elegant
ways to deal with this in future.