Implement Python3 support for gambit module.
This updates the gambit Python module implementation to support Python3.
All tests pass under both Python 2.7.15 and 3.7.2.

Most of the heavy lifting is in string handling, including in sometimes
unexpected places, because rational numbers are mediated via string
representations.  Better use of Cython's support for round-trips
between Python strings and std::string is made, which also simplifies
and clarifies the code in some places.

It is probable that housekeeping improvements can further be made
in places, including on exception generation for unexpected type arguments.