Bug fix/workaround in nfg logit tracing for clang on Mac
In clang 10.0.0 on Mac, the logit tracer was segfaulting.
This was traced to an incorrect reference to the game being
stored in the callback function.  Changing this from
const Game & to Game eliminated the bug.  As in any event
a Game is itself a pointer under the hood, there is no performance
implication for doing this.  The previous code appears to
be correct as written, and therefore a compiler bug seems to
be implicated.