Rationalise organisation of files.
This rationalises the organisation of source files, which should cure
a host of ills which have developed over the last few years.

In recent years it has become clear that the ability to compile
Gambit as a C++ shared library is not particularly useful.
Some of the layout of the files (and the inclusion of header file lists
in Makefile.am) had been done with this intention.

Instead, the model is that the library is accessed through several
types of drivers.  These include the command-line tools, and the
graphical interface.

In addition, the thought had been that the Python extension would
use the shared library.  This however is definitely not the way to
go for the purposes of packaging, especially vis-a-vis PyPI.
It does raise the small possibility of differences in behaviour
coming from building as part of the Python extension as opposed to
the other tools - but in truth this is no different than the risk
of some platform-based differences in behaviour.

With this commit, building with Python 3.8 should now be working.
Tests for other versions of Python, and for playing nice with pip,
are on the to-do list.