Reorganise Python files to be able to create sdist packages
This commit rearranges the file structure so `python setup.py sdist`
works as expected, and produces a self-contained collection of files
that can be built on their own.

As part of this, the extension module has been renamed to `pygambit`
from `gambit`, as `gambit` is already used on PyPI by an unrelated
(and apparently abandoned) project.