Standardise encoding of game savefiles, communication with external tools.
This cleans up some legacy issues pertaining to Unicode handling in
Python 2 versus 3.

(1) Game save files as returned by .write are now str, not bytes.
(2) Writing a game save file with an unknown format type raises an
    exception (rather than defaulting quietly to 'native')
(3) Communication with external tools is now properly encoded
    (as utf-8).

As a change related to (1), lxml, which is used in reading and writing
GTE format files, is now added to the list of dependencies.