Replace non-portable calls to bcmp() with memcmp()
There were a few legacy calls to bcmp() included in the original
gametracer implementation from ~20 years ago.  This commit replaces
them with the more standard memcmp().

This eliminates the need to check for bcmp() in configure, and also
ensures that the code being used for the C++-based tools and the
Python extension is the same (without having to implement the
bcmp check for Python builds...)