Correct handling of non-existent -d option to gambit-simpdiv
The gambit-simpdiv tool does not (yet) implement a -d option, yet
the options parser indicated it did.  As a result passing -d to the
tool would result in a call to abort(), rather than a sensible error message.

This fixes this by removing -d from the option parser, and thereby
fixes issue #187.