Introduce different smart-pointer approach for games.
The GameObject class only deletes a game element if it is marked as not
valid.  However, games are never invalid - so therefore games were never
getting deleted.

This introduces a BaseGameRep, which changes the deletion behaviour while
otherwise being compatible with the GameObjectPtr.

This is a needed but temporary solution; the plan is soon to migrate
the game representations to the STL smart pointers.