Probabilities at chance information sets must all sum to one.
This changes the way we set action probabilities at chance information sets.
We now require all probabilities to be set simultaneously, and enforce
that they must be non-negative and sum to one.

This withdraws the possibility of setting any probability individually,
without a deprecation period.  Although this is a breaking change, we view
the possibility of having an invalid probability distribution at a chance
node as a bug. Specifically, methods for computing Nash equilibria do not
check or normalise the distribution at chance nodes, and so misleading
output and difficult-to-detect errors result from improper distributions.

This closes #303.