Remove StrategicRestriction from pygambit.
This removes the experimental StrategicRestriction concept.  In
retrospect this was ill-advised, as this method of capturing that
two games were related was cumbersome and had poorly-defined
semantics.

Instead, we are replacing this with the ability to restrict a game
by making a deep copy using selected strategies only.  For the moment
it is left to calling code to decide how to relate games.