Bugfix: Comparing mixed strategy/behavior profiles
The pygambit of implementation of comparison between mixed strategy profiles or mixed behavior profiles was implemented as a lexicographic comparison between flattened lists.  This was wrong on both counts (the lexicographic comparison and the flattening).

This replaces comparisons with only equality and inequality, with equality requiring the RHS be a profile of the same type on the same game.