Migrate operations that change games to Game class.
This migrates most of the operations that result in a change to the structure of the game from being members of game elements to being members of the Game itself.

For the most part, old versions of these operations have been retained, with deprecation warnings issued as FutureWarning.  These are expected to be removed in 16.2.0.