Replace Game.num_nodes with Game.nodes
This removes Game.num_nodes, which was not well-motivated, and replaces it with
Game.nodes.

Game.nodes currently returns a list of nodes.  In future this might be better suited for a generator; the current implementation returns
a list to avoid issues with mutability and simultaneous/atomic application of edits.