In pygambit, access outcome payoffs only by player reference.
This changes the method for accessing outcome payoffs in pygambit from integer indices
to player references (either instances of Player, or label strings).
This aligns with the conventions used elsewhere in the package.

Alongside this, the user guide has been revised to focus on .from_arrays as the most natural quick-start way
to create a new game in strategic form.