Migrate from using nose2 to pytest.
With this we have moved from using nose2 to pytest for running the pygambit test suite.

Many, though not all, of the tests have been migrated from unittest to pytest style.

At present the remaining tests in unittest style are supported by pytest, but it will be desirable to rewrite those as well.