This removes gambit-enumpoly from the default build and marks as experimental.
The implementation of gambit-enumpoly has a number of problems which have cumulated over the years, possibly related to building pelican on 64-bit systems.

The roadmap calls for this method to be refactored heavily to separate out the concept of support enumeration from solution method, and to allow for different solution backends for given candidate supports.

As such, for now, we are doing the following:

1.  This method is now marked as "experimental" and removed from the build by default.
2.  References to it in the graphical interface have been removed.  (It was never the "recommended" solver so this has little effect unless someone was selecting it explicitly.)