Implement versioning of games and corresponding invalidation of mixed profiles.
Because games are mutable, MixedStrategyProfile and MixedBehaviorProfile objects can become invalid
due to changes in the underlying game.

This implements a mechanism at the C++ level which keeps track of the "version" of the game at the time that the profile is created,
and raises an exception if the underlying game has changed.