Fix bug in formulating LP tableau for sequence form.
This corrects a bug in formulating the LP tableau for the sequence form.
Specifically, there was a missing parenthesis in setting tableau entries; the effect of this is that
the tableau was incorrect in the case of a game which had a chance node that was not the
root node.

This closes #134