Standardise interfaces for collections in game representation.
This implements a standard convention for accessing collections of game elements.
Where appropriate, this provides GetXXXs() functions which, at present,
return Arrays of the corresponding GameObjectPtr.
Similar functions are provided for support profiles.

With these changes, iterator-type access for game members is now possible consistently across
using code, and this is preferred style.