Simplify and standardise sequence form implementation in LCP/LP.
This removes the obsolete uses of behavior support profiles from the implementations
of LCP and LP on the sequence form.
In addition, between the two the structure of populating the constraint tableau and recovering behavior strategies has been made parallel,
and parts re-expressed to clarify the intent of the code using more modern C++ idioms.