Special-case scaling logic for GNM for games with constant payoff
GNM implements payoff scaling that normalises payoffs in the general case to be between [0, 1].
This is a numerical convenience that isn't strictly necessary for the correctness of the method.
It's also a problem in the degenerate case where all of the payoffs in the game are the same.
Even though this is not generally considered an interesting case, the algorithm should behave
correctly when passed such a game.
This commit disables the payoff scaling in only this case.

Closes #405.