Check for non-zero perturbation vector in GNM.
GNM requires an initial vector along which to perturb the game, such that,
when this perturbation is scaled to be large enough, the game has a unique equilibrium.
This requires that the perturbation vector is nonzero.  This commit
adds a check to ensure this is the case,
and raises an exception if not.

Closes #194.