Check that perturbation vector for IPA is generic.
This adds a check that the perturbation vector specified for IPA
is generic.  Specifically, it requires that the perturbation
vector has a unique maximiser for each player.  This ensures
that the game perturbed along this ray has a unique equilibrium
in pure strategies for a sufficiently large perturbation.

This also changes the perturbation vector used by the default call to IPA
to perturb the game such that the perturbed game's equilibrium is
for each player to play their first-numbered strategy.

A consequence of using non-generic perturbation vectors was that
IPA could return non-Nash output.

Closes #406.