Remove Pelican implementation
The distribution has contained Pelican by Birk Huber as part of the suite for
solving polynomial systems.  Use of this was actually disabled in 2000 (commit 1cee1c53), and neither reinstated or removed.

Pelican is not 64-bit compatible and would take a huge effort to make it so.  Meanwhile the Gambit-native implementation
for solving polynomial systems appears completely viable,
so retaining Pelican is no longer necessary.