Expose regret-based stopping criterion for simpdiv_solve
Internally, the stopping criterion for simpdiv_solve has always been based on what we call the
maximum regret by any player.  This now exposes this as a parameter to simpdiv_solve in pygambit
and as a command-line parameter to the gambit-simpdiv tool.

The acceptance criterion is scaled by the range of payoffs in the game, so multiplying all payoffs in a game
by a common factor will result in the same output given the same acceptance tolerance.