New interfaces for generating random profiles.
C++:

This removes the .Randomize() members of `MixedStrategyProfile` and `MixedBehaviorProfile` and replaces them with
functions defined on `Game`.  The new functions use the STL random library
for generating randomisation, and require calling code to supply a random number
generator object.

Python:

This removes the .randomize() members of `MixedStrategyProfile` and `MixedBehaviorProfile` and replaces them with
`Game.random_strategy_profile` and `Game.random_behavior_profile`.

The new versions do not use the randomisation in the C++ classes, but instead use the `numpy`/`scipy`
facilities by accepting a generator as an optional argument, to facilitate reproducible
generation of profiles.

Both implementations correct a bug in which profiles were not truly uniform when
generated on a specified grid of rational probabilities.

Closes #295.