Use perturbations in gambit-logit to avoid infinite loops
The previous behaviour of the numerical continuation used to trace the logit QRE correspondence
attempted to deal with simple bifurcations by accepting the reversal of the orientation of the curve
and continuing "straight through".  However, in some cases, depending on where the tracing landed,
this could lead to an infinite loop (where the tracing repeatedly jumped back and forth between
two branches), or reversing the tracing entirely (and ending up back at the starting point).

Instead, the continuation now uses a perturbation approach and attempts to follow a path which has the same orientation.

It is still possible for the tracing to get stuck if it is too close to a bifurcation point.
This results in early termination of the tracing but not an infinite loop, and in the event this occurs
a limiting QRE is not marked as an epsilon-equilibrium.

Closes #3.