Refactor implementation of path-following and QRE.
This is a substantial refactoring and cleanup of path-following and the
interface to QRE tracing.

The only new functionality is to implement the ability to stop at an interior
local maximiser when fitting QRE.

Otherwise, other changes are internal to simplify calling the various
QRE-related routines and separate concerns (e.g. formatting/printing is no longer
part of the QRE tracer itself but properly delegated to a
provided observer function if that is desired).