Rework sequence form implementation in enumpoly.
This is a substantial reworking of the implementation of the sequence form concept
currently inside enumpoly.  This significantly simplifies the implementation
through a combination of better organisation and the use of
modern C++ features which did not exist when this implementation was
first done.

In due course a more polished version of this should become part of the
game-representation library.  This is progress towards that, but is not
yet fully stabilised.