Modernise further the implementation of enumpoly on sequence forms
This builds on previous work to bring the implementation of enumpoly on
extensive forms (via the sequence form) up-to-date.

This is a substantial refactoring that should improve the readability and
maintainability of these routines.

Further, the `GameSequenceForm` class has now moved towards being
ready to move to be part of the game representation library, which the goal
is to do in due course.