Debridement and modernisation of behavextend
This does some cleanup and updating of behavextend.  This involves no
functionality changes, but removes unneeded/duplicate code,
and uses modern features like STL containers and range-based
for-loops where it is straightforward to do so.

This is not a complete modernisation as there are a few places where
modernisation is not just a trivial reworking.

In addition, it is not at all clear that these functions are necessarily fully correct,
or that the task they are solving could not be handled more easily!