Reimplement PNS support enumeration algorithm
This reimplements our support enumeration algorithm for strategic games,
using the approach proposed by Porter, Nudelman and Shoham (2004).
An experimental version of this had been previously included, written by
Litao Wei.
This version is re-built from the ground up as a replacement.
Further, we now use this approach as the approach for generating supports,
and have removed the old method entirely, as there is no real
benefit to the way we used to do it over what PNS propose.