Enforce expected number of outcomes or payoffs in .nfg files.
Parsing of .nfg files has heretofore been somewhat tolerant.  If too many
payoffs or outcomes were given, extra ones were ignored silently;
if not enough were given, the missing ones were set to zero.

With this, checks are now done to ensure the correct number of outcomes or payoffs (as appropriate) are
provided.

Closes #119.