Refactor solve-by-subgames implementation.
This re-writes the implementation of solving by subgames.  This feature
is not yet widely visible.  However, this accomplishes two intermediate goals:

* Makes the implementation much more understandable;
* The implementation uses standard C++ containers instead of Gambit's.

There are plans for better ways to accomplish partial solution of games which will
subsume this.  Meanwhile, this will make the current implementation
more accessible and maintainable, and will permit it
to be made visible as appropriate to users.