Raise informative exception when trying to call `Game()`.
Games are intended to be created using one of the provided factory functions.
However, calling `Game()` did not immediately raise an exception; only when
attempting to operate on the game did a (less informative) exception occur.

This now raises a `ValueError` when calling the `Game` constructor.

Relatedly, this applies the technique across all game-related objects, and
standardises the way in which those objects are initialised internally.

Closes #463.