Migrate List to be a thin wrapper of std::list
This reimplements the Gambit List class to be a thin wrapper around
STL's std::list.  The only remaining non-STL operation is to be able
to index using the [] operator, which is 1-based.

Future new code should use std::list directly, and existing code using
List because of [] should be refactored to be able to use
std::list.