Make Array as similar to std::vector as possible.
This aligns the interface of Array to be as similar to std::vector as
possible, without (yet) adopting a std::vector implementation.

The indexing mechanism has been clarified, with first_index() and
last_index() methods which will be more compatible with this
being viewable as an extension of std::vector for the use case
of needing a contiguous set of indices not necessarily starting
at zero.